Improved dylib build script: no longer deletes build if going to fail, checks for CMake dependency, builds without external console bug

+++ LiteHtmlLib/build.sh
@@ -1,5 +1,10 @@
+if ! which "/usr/local/bin/cmake" >/dev/null ;  then
+    echo "CMake must be installed to build LiteHtml.dylib"
+    exit 2
+fi
+
 rm -rf build
 mkdir build
 cd build
-cmake ..
+/usr/local/bin/cmake ..
 make -j8
